backend/engines/dns_intelligence.py
"""
DNS Intelligence Engine - Advanced DNS Analysis and Intelligence
===============================================================

Provides comprehensive DNS intelligence including:
- Historical DNS record analysis
- DNS zone transfer testing
- Reverse DNS lookups
- DNS cache poisoning detection
- Domain reputation analysis
"""

import asyncio
import logging
import dns.resolver
import dns.reversename
from datetime import datetime
from typing import Dict, List, Optional
from dataclasses import dataclass

logger = logging.getLogger(__name__)

@dataclass
class DNSIntelligence:
    """DNS intelligence data"""
    domain: str
    record_type: str
    value: str
    ttl: int
    timestamp: datetime

class DNSIntelligenceEngine:
    """
    Advanced DNS intelligence gathering and analysis
    """

    def __init__(self):
        self.resolver = dns.resolver.Resolver()
        logger.info("DNS Intelligence Engine initialized")

    async def gather_dns_intelligence(self, target: str) -> List[Dict]:
        """Gather comprehensive DNS intelligence"""
        dns_records = []

        try:
            # Standard DNS records
            record_types = ['A', 'AAAA', 'CNAME', 'MX', 'NS', 'TXT', 'SOA']

            for record_type in record_types:
                try:
                    answers = self.resolver.resolve(target, record_type)
                    for answer in answers:
                        dns_records.append({
                            'domain': target,
                            'type': record_type,
                            'value': str(answer),
                            'ttl': answers.rrset.ttl,
                            'timestamp': datetime.now().isoformat()
                        })
                except Exception:
                    pass

            logger.info(f"Gathered {len(dns_records)} DNS records for {target}")

        except Exception as e:
            logger.error(f"DNS intelligence gathering failed: {str(e)}")

        return dns_records

logger.info("DNS Intelligence Engine module loaded")