backend/requirements/base.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database - PostgreSQL
sqlalchemy[asyncio]==2.0.23
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.0

# Cache & Queue
redis==4.6.0
celery[redis]==5.3.4

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
bcrypt==4.1.2
slowapi==0.1.9  # API rate limiting

# HTTP Requests
httpx==0.25.2
aiohttp==3.9.1
requests==2.31.0

# Scanning & Network Tools
python-nmap==0.7.1
dnspython==2.4.2
tld==0.13

# Data Processing
pandas==2.1.4
numpy==1.26.2

# Async & Concurrency
asyncio-mqtt==0.16.1

# Utilities
python-dateutil==2.8.2
email-validator==2.1.0
Pillow==10.1.0

# API Security Scanner
beautifulsoup4==4.12.2
lxml==4.9.3
playwright==1.40.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1